Optimize Test Execution and Add Jest Configurations
This commit enhances the efficiency and performance of the test suite
in our TypeScript starter project. The updates focus on bolstering the
test infrastructure, ensuring smoother and faster execution, and improving
the overall quality of the codebase.

The key changes in this commit include the addition of a Jest
configuration file and the inclusion of 'ts-jest' in the development
dependencies. The new Jest configuration file sets the test environment
to 'node' and specifies 'ts-jest' as the preset, paving the way for more
 effective testing of TypeScript files.

The math.test.ts file has also been adjusted to import the add
function from the math.ts file, enabling a more accurate representation
of module dependencies in our tests.

Lastly, the newTsp_install_deps.sh script has been updated to include
'ts-jest' in the installed development dependencies. This is a crucial
upgrade, as 'ts-jest' allows Jest to transpile TypeScript to JavaScript
before running the tests, contributing to more efficient and accurate testing.

In sum, these enhancements are designed to deliver a more streamlined,
 efficient, and accurate testing process, thereby bolstering the overall

uuid-stamp: 531bb738-9faa-469b-8341-8f72912e21f0