Refine package.json and TypeScript Configurations
This commit incorporates vital updates to the project structure and
TypeScript configurations to optimize the project's functionality.

The changes in this commit include the deletion of the
newTsp_init_git.sh file under development/typescripStarter/util.
This file was removed as the necessary git initializations are now
handled differently, contributing to the overall efficiency of the project setup.

In the newTsp_init_npm.sh file, npm init is implemented with
additional configurations added to the package.json file.
This enhancement ensures that the main file, types, type, and files
directory are appropriately set for the project.

Lastly, in the newTsp_init_typescript.sh file, TypeScript is
initialized with a particular set of parameters. Notably, the
strict mode is set to false in this update, providing flexibility
during the project development process.

The implemented modifications are essential for the efficient
functioning and advancement of the project, while also improving the
development experience.

uuid-stamp: 14e164d6-c9f5-4eec-b6ae-4034663cac86